requirements.txt
discord.py==2.6.3
fuzzywuzzy==0.18.0
mysql-connector==2.2.9
python-dotenv==1.1.1
requests==2.32.5
Unidecode==1.4.0
Levenshtein==0.27.1